colors in shared file
